requirements.txt
Django
djangorestframework
PyPDF2
python-dotenv
drf-yasg
